Check if process still exists when killing
 - renamed `paused` to `killSelected`
 - removed unnecessary once statement in updateProcs
 - if a process is selected for killing, check if the
   PID still exists on every update

Fixed selection color not persisting

+++ src/display/process/procGraphs.go
@@ -54,7 +54,7 @@ func getChildProcs(proc *process.Process) []string {
 // ProcVisuals renders graphs and charts for per-process stats.
 func ProcVisuals(ctx context.Context,
 	dataChannel chan *process.Process,
-	refreshRate int32) error {
+	refreshRate uint64) error {
 
 	if err := ui.Init(); err != nil {
 		log.Fatalf("failed to initialize termui: %v", err)
